linear_regression/requirements.txt
future==0.18.2
neuralpy-torch==0.0.3
numpy==1.18.4
pkg-resources==0.0.0
torch==1.5.0
